# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server to answer Stock related queries",
        "title": "Go Finance Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/forex": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Fetch forex rates of 165 currencies against a base currency at a given day from 1949",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forex"
                ],
                "summary": "Get forex rates of currencies against a base currency at a given day",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base Currency ",
                        "name": "base_currency",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date to fetch forex rates at",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/frankfurter.Rate"
                            }
                        }
                    }
                }
            }
        },
        "/price": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Fetch closing price of a company on any past date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Get price of a stock at a given day",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company Name ",
                        "name": "company",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date to fetch price at",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.StockResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "frankfurter.Rate": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "quote": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                }
            }
        },
        "handlers.StockResponse": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}